pyproject.toml
[tool.poetry]
name = "agwise-potato"
version = "1.0.0"
description = "API to fetch potato planting data based on several filter prarameters"
authors = ["Sammy Barasa <barsamms@gmail.com>"]
readme = "README.md"
packages = [{include = "agwise_potato"}]

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
pymysql = "^1.1.0"
#typer = {extras = ["all"], version = "^0.9.0"}
colorlog = "^6.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
